# specs/main/contracts/llm-platform.openapi.yaml
openapi: 3.1.0
info:
  title: LLM Platform API
  version: 0.1.0
  description: >
    Core platform contract for tool registration, agent registration, invoke,
    file access, ACL management, and trace query.
servers:
  - url: http://127.0.0.1:3000
paths:
...
  /v1/files:
    get:
      summary: Read file or list direct children
      operationId: getFile
      parameters:
        - name: path
          in: query
          schema:
            type: string
      responses:
        '200':
          description: File or directory payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileReadResponse'
    put:
      summary: Write file
      operationId: putFile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path, content]
              properties:
                path:
                  type: string
                content:
                  type: string
      responses:
        '200':
          description: File saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileWriteResponse'
    delete:
      summary: Delete file or empty directory
      operationId: deleteFile
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Path deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileDeleteResponse'
  /v1/files/tree:
    get:
      summary: List file tree
      operationId: getFileTree
      parameters:
        - name: path
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Tree payload
          content:
            application/json:
              schema:
                type: object
                required: [status, rootPath, items]
                properties:
                  status:
                    type: string
                    const: success
                  rootPath:
                    type: string
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/FileTreeItem'
  /v1/files/acl:
    get:
      summary: List ACL rules
      operationId: listFileAcl
      parameters:
        - name: userId
          in: query
          schema:
            type: string
        - name: path
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ACL rules
          content:
            application/json:
              schema:
                type: object
                required: [status, rules]
                properties:
                  status:
                    type: string
                    const: success
                  rules:
                    type: array
                    items:
                      $ref: '#/components/schemas/FileAclRule'
    post:
      summary: Upsert ACL rule
      operationId: upsertFileAcl
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileAclRule'
      responses:
        '200':
          description: ACL rule stored
          content:
            application/json:
              schema:
                type: object
                required: [status, rule]
                properties:
                  status:
                    type: string
                    const: success
                  rule:
                    $ref: '#/components/schemas/FileAclRule'
    delete:
      summary: Delete ACL rule
      operationId: deleteFileAcl
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ACL rule deleted
          content:
            application/json:
              schema:
                type: object
                required: [status, deleted]
                properties:
                  status:
                    type: string
                    const: success
                  deleted:
                    type: boolean
...
components:
...
  schemas:
...
    FileTreeItem:
      type: object
      required: [path, type]
      properties:
        path:
          type: string
        type:
          type: string
          enum: [file, dir]
    FileReadResponse:
      type: object
      required: [status, path, type]
      properties:
        status:
          type: string
          const: success
        path:
          type: string
        type:
          type: string
          enum: [file, dir]
        content:
          type: string
        items:
          type: array
          items:
            type: object
            additionalProperties: true
    FileWriteResponse:
      type: object
      required: [status, path, saved]
      properties:
        status:
          type: string
          const: success
        path:
          type: string
        type:
          type: string
        saved:
          type: boolean
    FileDeleteResponse:
      type: object
      required: [status, path, deleted]
      properties:
        status:
          type: string
          const: success
        path:
          type: string
        type:
          type: string
        deleted:
          type: boolean
    FileAclRule:
      type: object
      required: [userId, path, permissions]
      properties:
        userId:
          type: string
        path:
          type: string
        permissions:
          type: array
          items:
            type: string
            enum: [read, write, delete]
...